+++ test/exported/error.test.ts
@@ -7,6 +7,7 @@ import {
   InvalidConditionError,
   InvalidLimitValueError,
   InvalidOffsetValueError,
+  InsertColumnsAndValuesNotEqualError,
   Builder,
   e,
   Table,
@@ -28,6 +29,7 @@ const ADD = ArithmeticOperator.ADD
 const GT = ComparisonOperator.GreaterThan
 const table1 = database.s.public.t.table1
 const col1 = table1.c.col1
+const col2 = table1.c.col2
 const col3 = table1.c.col3
 const col4 = table1.c.col4
 const table2 = database.s.public.t.table2
@@ -319,6 +321,31 @@ describe('Throw desired Errors', () => {
       }
 
       expect(actual).toThrow(/^Aggregate function SUM cannot be used in RETURNING clause$/)
+    })
+  })
+  describe('Error: InsertColumnsAndValuesNotEqualError', () => {
+    it(`columns more than values`, () => {
+      function actual() {
+        sql.insertInto(table1, col1, col2).values('A')
+      }
+
+      expect(actual).toThrow(InsertColumnsAndValuesNotEqualError)
+    })
+    it(`values more than columns`, () => {
+      function actual() {
+        sql.insertInto(table1, col1, col2).values('A', 'B', 'C')
+      }
+
+      expect(actual).toThrow(InsertColumnsAndValuesNotEqualError)
+    })
+    it(`values contain unsupported type`, () => {
+      const value = { unsupportedObject: 'something' }
+      function actual() {
+        // @ts-ignore
+        sql.insertInto(table1).values(value).getSQL()
+      }
+
+      expect(actual).toThrow(`Value step has Unsupported value: ${value}, type: ${typeof value}`)
     })
   })
 })